config-web-ui/requirements.txt
Flask
ruamel.yaml
docker
requests